Moved fonts into a constant file
