+++ apps/web/src/utils/constants.ts
@@ -1,0 +1,6 @@
+export const FONTS = {
+  arial: 'Arial, sans-serif',
+  timesNewRoman: 'Times New Roman, serif',
+  helvetica: 'Helvetica, sans-serif',
+  segoeUI: 'Segoe UI, Tahoma, Geneva, Verdana, sans-serif',
+} as const;
